Fix: Fixed navbar slide-out for small screens
Current implmentation of this fix might be sort of hackish. I know it can be improved with the usage of await & promises
but that isn't the issue. This code is becoming more complex than it needs to be. Perhaps, saving markup wasn't that good of an idea.
We wouldn't need this piece of code if our mobile navbar was separate from the header

JS code will have to be organized, there is a lot of redundancy & repetition